show default data by default
